Fixed list sort on images, networks, volumes
